- rename RepoClient to ContentApi - uncomment code - reorg files structure
